requirements.txt
Django==5.2 
djangorestframework==3.16.0 
django-cors-headers==4.7.0
openai==1.75.0
python-dotenv==1.1.0
mcp==1.6.0
httpx
starlette
sse-starlette
uvicorn